I have done private fields and deleted white space, moved methods to the other classes
