+++ src/main/java/DataModel/ExchangeRate.java
@@ -10,35 +10,9 @@ public class ExchangeRate {
     private String ticker;
     private String date;
     private double exchangeAmount;
-    private FetchToTheForexAPI fetch = new FetchToTheForexAPI();
-    private DataFromForexApi dataFrom = new DataFromForexApi();
-    private JSONMapper mapper = new JSONMapper();
     private double rate;
 
-    public double getExchangeRateToBeConverted(String date) {
-        setFetchToTheForexApi(date);
-        HashMap maps = (HashMap) dataFrom.getRates();
-        return (double) maps.get("PLN");
-    }
-
-    public double getExchangeRate(String date) {
-        setFetchToTheForexApi(date);
-        HashMap maps = (HashMap) dataFrom.getRates();
-        return roundTo2DecimalPlace((double) maps.get("PLN"));
-    }
-
-    public void setFetchToTheForexApi(String date) {
-        if (Objects.equals(date, null)) {
-            String fetchInfo = fetch.getRespondsWithLatestDate();
-            dataFrom = mapper.mapJsonToJava(fetchInfo);
-        } else {
-            String fetchInfo = fetch.getRespondsWithHistoricalDate(date);
-            dataFrom = mapper.mapJsonToJava(fetchInfo);
-        }
-    }
-
     public ExchangeRate() {
-
     }
 
     public ExchangeRate(String ticker, double exchangeAmount, double rate) {
@@ -77,20 +51,6 @@ public class ExchangeRate {
 
     public void setRate(double rate) {
         this.rate = rate;
-    }
-
-    public double convertedCurrency(double exchangeAmount, String date) {
-        if (exchangeAmount <= 0) {
-            System.out.println("You have entered an incorrect amount. The amount cannot be negative or zero.");
-            return 0;
-        } else {
-            double result = exchangeAmount * getExchangeRateToBeConverted(date); //exchangeAmount
-            return roundTo2DecimalPlace(result);
-        }
-    }
-
-    private double roundTo2DecimalPlace(double value) {
-        return Math.round(value * 100.0) / 100.0;
     }
 
     @Override
